refactor(tests): move MockRequests class to a separate mocks module
